Updates to work with new GHC versions (8.6.5)
1. Package dependencies
- parsec now uses version >= 3 instead of >= 2.7 < 3
- uniplate now uses version >= 1.6.1 instead of == 1.6.1
- template-haskell now uses version >= 2.11 < 3 instead of >= 2.7 < 3 (motivated by signature change in DataD)

According to `cabal outdated` no other packages have to be updated.

2. `Applicative` superclass of `Monad`
`Monad`s without existing `Applicative` instances are now defined as
```
(<*>) = ap
pure = return
```
Affected types:
- PiReader

3. `MonadFail`
Adopted monad instances to work with the new `MonadFail` according to recommendations given in the proposal: https://prime.haskell.org/wiki/Libraries/Proposals/MonadFail#Adaptingoldcode

The old `fail` instances of the `Monad`class have been moved to the `MonadFail` class. The `fail` of the `Monad` class is now defined to be equal to the `MonadFail` one.

Affected types:
- PiReader
- TcCodeM
- TcDeclM
- InterpretM. There was no custom definition of `fail` so it is now defined as `fail s = error s` (rather than the old default `fail s = errorWithoutStackTrace s`).

Furthermore the class `HasSubTyping` now has `MonadFail` as a superclass. It is possible that this is overkill but it seems to work fine.

4. The `<>` of `Semigroup`
The `<>` operator is now a exported by `Prelude` and collides with existing definitions. It has been solved by not importing `<>` from `Prelude`.

5. `Typeable` instances in .hs-boot files were not accepted by the compiler. These have been commented out and it is unknown whether this as had any effect.

6. `DataD` and `NewtypeD` signatures change (from template-haskell). Updated pattern matches with `_` for the new argument.

7. Removed `Applicative` instance for `GenParser s a`, one is already provided.

8. Flowlocks updates. The `MonadFail` problem occurred in few places in the ~flowlocks-framework~. These changes are not yet on hackage, so in the mean-time a copy of the updated framework needs to be manually kept in the project root.